refactor: atualizando base e orm
