--- api-domrock/src/main/java/com/dataflow/apidomrock/entities/database/Status.java
@@ -1,20 +1,0 @@
-package com.dataflow.apidomrock.entities.database;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "status")
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class Status {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String status;
-}
