+++ api-domrock/src/main/java/com/dataflow/apidomrock/entities/enums/NivelAcessoEnum.java
@@ -1,4 +1,4 @@
-package com.dataflow.apidomrock.services.utils;
+package com.dataflow.apidomrock.entities.enums;
 
 public enum Niveis {
     MASTER,
